Add files via upload
Stocks are an attractive investment option but the risk can also result in large losses in a short time. Thus, it requires careful attention to stock price movement and accurate data modeling is needed to forecast stock prices with a low error rate.Forecasting models using deep learning are believed to be
able to predict stock price movements accurately with time-series data input, especially the Long Short-Term Memory(LSTM) and Gated Recurrent Unit (GRU) algorithms

This article aims to build models using Recurrent Neural Networks (RNN),  Long-Short Term Memory model (LSTM) and Gated Recurrent Units (GRU) to predict future stock market price. The main objective of this paper is to experiment each of the model.

It is found that the best model is GRU model with sequence length of 100 and learning 0.001. The test result of the the best model is RMSE of 23.52 and MAE of 21.32. It is found that the prediction of the test data is generally lower then the actual stock price.